requirements.txt
# Apache Airflow
apache-airflow==2.7.3
apache-airflow-providers-postgres==4.3.0

# DBT 
dbt-core==1.9.3
dbt-postgres==1.9.0

# PostgreSQL Database Connection
psycopg2-binary==2.9.9
sqlalchemy==2.0.29

# Environment Variables Handling
python-dotenv==1.0.1

# Data Processing
pandas==2.2.1
numpy==1.26.4



pyyaml==6.0.1
tabulate==0.9.0

